# docs/swagger.yaml
definitions:
...
  dto.PlatformItem:
    description: A delivery platform (Blinkit, Zepto, ...). qc_name maps it onto
    properties:
      code:
        description: Short code used in requests, e.g. "blinkit"
        type: string
      created_at:
        description: Creation timestamp, RFC3339
        type: string
      delivery_eta_text:
        description: Fallback delivery ETA copy, e.g. "10 Mins"
        type: string
      display_name:
        description: Display name, e.g. "Blinkit"
        type: string
      display_order:
        description: Sort order in pickers
        type: integer
      enabled:
        description: Whether the platform is enabled
        type: boolean
      logo_url:
        description: URL of the platform's logo
        type: string
      platform_id:
        description: Unique identifier for the platform
        type: string
      qc_name:
        description: The platform's name on QuickCommerce, e.g. "BlinkIt". Empty =
          not mapped.
        type: string
      searchable:
        description: True when qc_name is set, i.e. the platform can be searched/linked.
        type: boolean
    type: object
...
  dto.Response:
    description: Generic response envelope wrapping every API response.
    properties:
      code:
        description: Machine-readable error/status code, present on failures
        type: string
      data:
        description: Response payload, shape depends on the endpoint
      message:
        description: Human-readable message describing the result
        type: string
      status:
        description: Outcome of the request, e.g. "success" or "failed"
        type: string
    type: object
...
  service.BatchLinkFailure:
    properties:
      error:
        type: string
      platform_code:
        type: string
      variant_id:
        type: string
    type: object
  service.BatchLinkResult:
    properties:
      failed:
        items:
          $ref: '#/definitions/service.BatchLinkFailure'
        type: array
      mapped:
        type: integer
    type: object
  service.Candidate:
    properties:
      available:
        type: boolean
      brand:
        type: string
      deeplink:
        type: string
      inventory:
        type: integer
      is_combo:
        type: boolean
      is_multipack:
        type: boolean
      mrp_paise:
        type: integer
      multipack:
        type: integer
      name:
        type: string
      out_of_stock:
        type: boolean
      price_paise:
        type: integer
      qc_item_id:
        type: string
      quantity:
        type: string
      stock_label:
        type: string
    type: object
  service.RefreshResult:
    properties:
      credits_remaining:
        type: integer
      failed:
        type: integer
      refreshed:
        type: integer
    type: object
...
  v1.BatchLinkEntry:
    properties:
      available:
        type: boolean
      deep_link:
        type: string
      inventory:
        type: integer
      mrp_paise:
        type: integer
      platform_code:
        type: string
      price_paise:
        type: integer
      qc_item_id:
        type: string
      variant_id:
        type: string
    required:
    - platform_code
    - qc_item_id
    - variant_id
    type: object
  v1.BatchLinkRequest:
    properties:
      city_id:
        type: string
      links:
        items:
          $ref: '#/definitions/v1.BatchLinkEntry'
        minItems: 1
        type: array
    required:
    - city_id
    - links
    type: object
  v1.ConfirmLinkRequest:
    properties:
      available:
        type: boolean
      city_id:
        type: string
      deep_link:
        type: string
      inventory:
        type: integer
      mrp_paise:
        type: integer
      platform_code:
        type: string
      price_paise:
        type: integer
      qc_item_id:
        type: string
    required:
    - city_id
    - platform_code
    - qc_item_id
    type: object
...
  v1.CreatePlatformRequest:
    properties:
      code:
        type: string
      delivery_eta_text:
        type: string
      display_name:
        type: string
      enabled:
        type: boolean
      logo_url:
        type: string
      qc_name:
        type: string
    required:
    - code
    - display_name
    type: object
...
  v1.CreditsResponse:
    properties:
      credits_remaining:
        type: integer
    type: object
...
  v1.DeletePlatformRequest:
    properties:
      platform_id:
        type: string
    required:
    - platform_id
    type: object
...
  v1.LinkSearchResponse:
    properties:
      credits_remaining:
        type: integer
      results:
        additionalProperties:
          items:
            $ref: '#/definitions/service.Candidate'
          type: array
        type: object
    type: object
...
  v1.ManualPriceRequest:
    properties:
      city_id:
        type: string
      deep_link:
        type: string
      mrp_paise:
        type: integer
      platform_code:
        type: string
      platform_sku:
        type: string
      price_paise:
        type: integer
      variant_id:
        type: string
    required:
    - city_id
    - mrp_paise
    - platform_code
    - price_paise
    - variant_id
    type: object
...
  v1.PlatformOption:
    properties:
      code:
        type: string
      display_name:
        type: string
      logo_url:
        type: string
      platform_id:
        type: string
      qc_name:
        type: string
      searchable:
        type: boolean
    type: object
...
  v1.RefreshRequest:
    properties:
      city_id:
        type: string
      product_id:
        type: string
      variant_id:
        type: string
    required:
    - city_id
    type: object
...
  v1.UpdatePlatformRequest:
    properties:
      code:
        type: string
      delivery_eta_text:
        type: string
      display_name:
        type: string
      enabled:
        type: boolean
      logo_url:
        type: string
      platform_id:
        type: string
      qc_name:
        type: string
    required:
    - platform_id
    type: object
...
  v1.VariantPriceRow:
    properties:
      available:
        type: boolean
      deep_link:
        type: string
      inventory:
        type: integer
      last_updated_at:
        type: string
      mrp_paise:
        description: MRP in paise. 0 when the platform reports no MRP.
        type: integer
      platform_code:
        type: string
      platform_id:
        type: string
      platform_name:
        type: string
      platform_sku:
        type: string
      price_paise:
        description: Selling price in paise. 3800 = ₹38.00
        type: integer
      source:
        type: string
    type: object
...
info:
  contact:
    name: API Support
  description: This is the API documentation for the Grocerics backend service.
  title: Grocerics API
  version: "1.1"
paths:
...
  /v1/credits:
    get:
      description: Live QuickCommerce balance, for the admin navbar. Asking is free
        — it costs no credit itself. Every search costs 1 credit per platform; linking
        costs 0 when the candidate is relayed, and the fan-out to other enabled cities
        costs 1 per item per city.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.CreditsResponse'
              type: object
      security:
      - BearerAuth: []
      summary: QuickCommerce credits remaining
      tags:
      - linking
...
  /v1/inventory-management/link/batch:
    post:
      consumes:
      - application/json
      description: 'Batch form of confirm-link: one search response yields every (variant,
        platform) mapping the admin picked, sent in one call. Send price_paise + available
        per entry (from the response you''re showing) and it costs 0 credits. Entries
        are independent — a bad entry is reported in `failed` and the rest still map,
        since every write is an idempotent upsert; re-send just the failures. Afterwards
        the backend fetches these same item ids in every OTHER enabled city in the
        background, so each city gets its own real price (1 credit per item per city).'
      parameters:
      - description: Batch link request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.BatchLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.BatchLinkResult'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Map many variants to platforms from one search response
      tags:
      - linking
  /v1/inventory-management/link/search:
    get:
      description: 'Discovery step: groupsearch across platforms so the admin can
        confirm the right item per platform. Costs 1 credit per platform. Never called
        on a consumer request.'
      parameters:
      - description: search term
        in: query
        name: q
        required: true
        type: string
      - description: city ID (supplies the QC location anchor)
        in: query
        name: city
        required: true
        type: string
      - description: comma-separated platform codes to search, e.g. blinkit,zepto.
          Omit to search every enabled QC-mapped platform. Costs 1 credit per platform.
        in: query
        name: platforms
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.LinkSearchResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Search QuickCommerce candidates for linking
      tags:
      - linking
  /v1/inventory-management/refresh:
    post:
      consumes:
      - application/json
      description: Re-pulls live price+stock via GetItem for a variant or a whole
        product's variants in one city. Costs 1 credit per linked platform. One platform
        failing never fails the refresh.
      parameters:
      - description: Refresh request (one of product_id / variant_id)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.RefreshResult'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Refresh live prices/stock from QuickCommerce
      tags:
      - linking
...
  /v1/inventory-management/variants/{variant_id}/links:
    post:
      consumes:
      - application/json
      description: Pins a (variant, platform) to the chosen QC item id and seeds its
        price/stock. Send price_paise + available from the search response you're
        already showing and this costs 0 credits; omit them and the server re-fetches
        via GetItem for 1 credit. The link is city-independent; the price is stored
        per city and is a reference only — mobile re-fetches live prices for the user's
        own pincode at cart time.
      parameters:
      - description: Variant ID
        in: path
        name: variant_id
        required: true
        type: string
      - description: Confirm link request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.ConfirmLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Confirm a platform link for a variant
      tags:
      - linking
  /v1/inventory-management/variants/{variant_id}/prices:
    get:
      description: Joins links + platform_prices for a variant in a city — powers
        the wizard's Variations grid.
      parameters:
      - description: Variant ID
        in: path
        name: variant_id
        required: true
        type: string
      - description: City ID
        in: query
        name: city
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/v1.VariantPriceRow'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Per-platform prices for a variant
      tags:
      - linking
  /v1/inventory-management/variants/manual-price:
    post:
      consumes:
      - application/json
      description: 'Admin fallback: pin the link and write the price by hand. price
        = MRP, discounted_price = shown price.'
      parameters:
      - description: Manual price request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.ManualPriceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Set a manual price (fallback when not on QuickCommerce)
      tags:
      - linking
...
  /v1/platforms:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Delete Platform Request
        in: body
        name: DeletePlatformRequest
        required: true
        schema:
          $ref: '#/definitions/v1.DeletePlatformRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete a platform
      tags:
      - platforms
    get:
      description: Enabled delivery platforms — powers the linking picker's platform
        checkboxes. Only those with searchable=true can be queried on QuickCommerce.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/v1.PlatformOption'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List enabled platforms
      tags:
      - linking
    patch:
      consumes:
      - application/json
      description: Set qc_name to make a platform searchable on QuickCommerce; clear
        it (send "-") to make it unsearchable.
      parameters:
      - description: Update Platform Request
        in: body
        name: platform
        required: true
        schema:
          $ref: '#/definitions/v1.UpdatePlatformRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PlatformItem'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update a platform
      tags:
      - platforms
    post:
      consumes:
      - application/json
      description: qc_name is the platform's name on QuickCommerce (e.g. "BlinkIt",
        "Zepto", "Swiggy"). Leave it blank and the platform exists but cannot be searched
        or linked.
      parameters:
      - description: Create Platform Request
        in: body
        name: platform
        required: true
        schema:
          $ref: '#/definitions/v1.CreatePlatformRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PlatformItem'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create a platform
      tags:
      - platforms
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"